pyproject.toml
[project]
name = "unsub"
version = "0.1.0"
description = "Unsubscribe from spam emails"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-auth",
    "google-auth-oauthlib",
    "google-api-python-client",
    "beautifulsoup4",
    "openai",
    "selenium",
    "Pillow",
]

[tool.setuptools]
packages = ["unsub", "unsub.simulations", "unsub.cmd"]

[tool.setuptools.package-data]
"unsub" = [
    "simulations/assets/**/*",
    "cmd/view_chat_assets/*",
]